restructure to be more d3-like
